# indexer/services/comlink/public/swagger.json
{
  "components": {
...
    "schemas": {
...
      "IsoString": {
        "type": "string"
      },
...
      "OrderSide": {
        "enum": [
          "BUY",
          "SELL"
        ],
        "type": "string"
      },
      "Liquidity": {
        "enum": [
          "TAKER",
          "MAKER"
        ],
        "type": "string"
      },
      "FillType": {
        "enum": [
          "LIMIT",
          "LIQUIDATED",
          "LIQUIDATION",
          "DELEVERAGED",
          "OFFSETTING"
        ],
        "type": "string"
      },
      "MarketType": {
        "enum": [
          "PERPETUAL",
          "SPOT"
        ],
        "type": "string"
      },
      "FillResponseObject": {
        "properties": {
          "id": {
            "type": "string"
          },
          "side": {
            "$ref": "#/components/schemas/OrderSide"
          },
          "liquidity": {
            "$ref": "#/components/schemas/Liquidity"
          },
          "type": {
            "$ref": "#/components/schemas/FillType"
          },
          "market": {
            "type": "string"
          },
          "marketType": {
            "$ref": "#/components/schemas/MarketType"
          },
          "price": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "createdAtHeight": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "clientMetadata": {
            "type": "string"
          },
          "subaccountNumber": {
            "type": "integer",
            "format": "int32"
          },
          "closedPnL": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "side",
          "liquidity",
          "type",
          "market",
          "marketType",
          "price",
          "size",
          "fee",
          "createdAt",
          "createdAtHeight",
          "subaccountNumber"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "FillResponse": {
        "properties": {
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalResults": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "fills": {
            "items": {
              "$ref": "#/components/schemas/FillResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "fills"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Indexer API",
    "version": "v1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/fills": {
      "get": {
        "operationId": "GetFills",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FillResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "marketType",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MarketType"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
    "/fills/parentSubaccount": {
      "get": {
        "operationId": "GetFillsForParentSubaccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FillResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "parentSubaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "marketType",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MarketType"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "",
      "description": "Public Testnet"
    }
  ]
}